refactor: move TriggerCreationModal path
