+++ Composer/packages/client/src/components/TriggerCreationModal/TriggerCreationModal.tsx
@@ -436,5 +436,3 @@ export const TriggerCreationModal: React.FC<TriggerCreationModalProps> = (props)
     </Dialog>
   );
 };
-
-export default TriggerCreationModal;
