+++ Composer/packages/client/src/components/TriggerCreationModal/index.tsx
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { TriggerCreationModal } from './TriggerCreationModal';
+export default TriggerCreationModal;
